requirements.txt
# Core dependencies
python-telegram-bot==22.6
google-generativeai==0.8.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.2.1
pydantic==2.11.7
pydantic-settings==2.12.0

# News fetching
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.32.5
aiohttp==3.11.12

# Database
sqlalchemy==2.0.46
alembic==1.18.3

# Scheduling
APScheduler==3.11.2

# Utilities
python-dateutil==2.9.0
pytz==2025.2
